+++ domain/src/main/java/net/boreeas/riotapi/constants/LevelUpType.java
@@ -14,17 +14,12 @@
  * limitations under the License.
  */
 
-package net.boreeas.riotapi.rest;
-
-import lombok.Getter;
+package net.boreeas.riotapi.constants;
 
 /**
  * @author Malte Schütze
  */
-@Getter
-public class ParticipantPlayer {
-    private String matchHistoryUri;
-    private int profileIconId;
-    private String summonerName;
-    private long summonerId;
+public enum LevelUpType {
+    NORMAL,
+    EVOLVE,
 }