+++ loginqueue/src/test/java/net/boreeas/riotapi/loginqueue/LoginQueueTest.java
@@ -20,7 +20,9 @@ import junit.framework.TestCase;
 import net.boreeas.riotapi.Shard;
 import net.boreeas.riotapi.com.riotgames.platform.account.management.InvalidCredentialsException;
 
-import java.net.URLEncoder;
+import java.io.FileInputStream;
+import java.io.InputStreamReader;
+import java.util.Properties;
 import java.util.concurrent.TimeUnit;
 
 public class LoginQueueTest extends TestCase {
@@ -34,7 +36,6 @@ public class LoginQueueTest extends TestCase {
     }
 
     public void testQueueWait() throws Exception {
-        System.out.println(URLEncoder.encode(Shard.NA.loginQueue, "UTF-8"));
         try {
             AuthResult result = new LoginQueue(Shard.NA).waitInQueue("foo", "").await(2, TimeUnit.SECONDS);
             fail();
@@ -48,5 +49,13 @@ public class LoginQueueTest extends TestCase {
             fail();
         } catch (IllegalStateException ex) {
         }
+    }
+
+    public void testQueueAuth() throws Exception {
+        Properties prop = new Properties();
+        prop.load(new InputStreamReader(new FileInputStream("testconfig.properties")));
+
+        String token = new LoginQueue(Shard.EUW).waitInQueueBlocking(prop.getProperty("user"), prop.getProperty("pass"));
+        System.out.println("LQ token: " + token);
     }
 }