+++ rest/src/main/java/net/boreeas/riotapi/rest/Match.java
@@ -17,6 +17,8 @@
 package net.boreeas.riotapi.rest;
 
 import lombok.Getter;
+import net.boreeas.riotapi.com.riotgames.platform.game.GameMode;
+import net.boreeas.riotapi.com.riotgames.platform.game.GameType;
 import net.boreeas.riotapi.com.riotgames.platform.game.QueueType;
 
 import java.util.ArrayList;
@@ -35,6 +37,8 @@ public class Match {
     private String matchVersion;
     private List<ParticipantIdentity> participantIdentities = new ArrayList<>();
     private List<Participant> participants = new ArrayList<>();
-    private QueueType queueType;
     private String region;
+    private GameMode matchMode;
+    private GameType matchType;
+    private QueueType queueType;
 }