+++ rest/src/main/java/net/boreeas/riotapi/rest/Rune.java
@@ -16,15 +16,10 @@
 
 package net.boreeas.riotapi.rest;
 
-import lombok.Getter;
-
 /**
  * @author Malte Schütze
  */
-@Getter
-public class ParticipantPlayer {
-    private String matchHistoryUri;
-    private int profileIconId;
-    private String summonerName;
-    private long summonerId;
+public class Rune {
+    private long id;
+    private long ranks;
 }