+++ rest/src/main/java/net/boreeas/riotapi/rest/ThrottledApiHandler.java
@@ -769,6 +769,33 @@ public class ThrottledApiHandler implements AutoCloseable {
 
     // </editor-fold>
 
+    // <editor-fold desc="Status">
+    /**
+     * <p>
+     * Retrieves general information about each shard.
+     * </p><br>
+     * This method does not count towards your rate limit.
+     * @return A list of shard infomation.
+     * @see <a href="https://developer.riotgames.com/api/methods#!/835/2939">Official API Documentation</a>
+     */
+    public Future<List<ShardData>> getShards() {
+        return new DummyFuture<>(handler.getShards());
+    }
+
+    /**
+     * <p>
+     * Retrieves detailed information about the specified shard.
+     * </p><br>
+     * This method does not count towards your rate limit.
+     * @param shard The target region
+     * @return A list of shard infomation.
+     * @see <a href="https://developer.riotgames.com/api/methods#!/835/2938">Official API Documentation</a>
+     */
+    public Future<ShardStatus> getShardSatatus(Shard shard) {
+        return new DummyFuture<>(handler.getShardStatus(shard));
+    }
+    // </editor-fold>
+
     // <editor-fold desc="Match">
     /**
      * Retrieves the specified match, including timeline.
@@ -802,6 +829,49 @@ public class ThrottledApiHandler implements AutoCloseable {
      */
     public Future<List<MatchSummary>> getMatchHistory(long playerId) {
         return new ApiFuture<>(() -> handler.getMatchHistory(playerId));
+    }
+
+    /**
+     * Retrieve a player's match history.
+     *
+     * @param playerId    The id of the player.
+     * @param championIds The championIds to use for retrieval.
+     * @return The match history of the player.
+     * @see <a href="https://developer.riotgames.com/api/methods#!/805/2847">Official API Documentation</a>
+     */
+    public Future<List<MatchSummary>> getMatchHistory(long playerId, String... championIds) {
+        return new ApiFuture<>(() -> handler.getMatchHistory(playerId, championIds));
+    }
+
+    /**
+     * Retrieve a player's match history, filtering out all games not in the specified queues.
+     *
+     * @param playerId    The id of the player.
+     * @param championIds The championIds to use for retrieval.
+     * @param queueTypes  The queue types to retrieve (must be one of RANKED_SOLO_5x5, RANKED_TEAM_3x3 or
+     *                    RANKED_TEAM_5x5).
+     * @return The match history of the player.
+     * @see <a href="https://developer.riotgames.com/api/methods#!/805/2847">Official API Documentation</a>
+     */
+    public Future<List<MatchSummary>> getMatchHistory(long playerId, String[] championIds, QueueType... queueTypes) {
+        return new ApiFuture<>(() -> handler.getMatchHistory(playerId, championIds, queueTypes));
+    }
+
+    /**
+     * Retrieve a player's match history, filtering out all games not in the specified queues.
+     * This method only returns games starting with beginIndex and ending at endIndex.
+     *
+     * @param playerId    The id of the player.
+     * @param championIds The championIds to use for retrieval.
+     * @param queueTypes  The queue types to retrieve (must be one of RANKED_SOLO_5x5, RANKED_TEAM_3x3 or
+     *                    RANKED_TEAM_5x5).
+     * @param beginIndex  The index of the first game that should be retrieved.
+     * @param endIndex    The index of the last game the should be retrieved.
+     * @return The match history of the player.
+     * @see <a href="https://developer.riotgames.com/api/methods#!/805/2847">Official API Documentation</a>
+     */
+    public Future<List<MatchSummary>> getMatchHistory(long playerId, String[] championIds, QueueType[] queueTypes, int beginIndex, int endIndex) {
+        return new ApiFuture<>(() -> handler.getMatchHistory(playerId, championIds, queueTypes, beginIndex, endIndex));
     }
     // </editor-fold>