+++ spectator/src/main/java/net/boreeas/riotapi/spectator/Callback.java
@@ -21,4 +21,10 @@ package net.boreeas.riotapi.spectator;
  */
 public @FunctionalInterface interface Callback {
     void receive();
+
+    default Callback and(Callback callback) {
+        if (callback == null) return this;
+
+        return () -> {receive(); callback.receive();};
+    }
 }