+++ spectator/src/main/java/net/boreeas/riotapi/spectator/GameUpdateTask.java
@@ -16,6 +16,7 @@
 
 package net.boreeas.riotapi.spectator;
 
+import lombok.Getter;
 import lombok.Setter;
 import lombok.extern.log4j.Log4j;
 import net.boreeas.riotapi.RequestException;
@@ -44,13 +45,15 @@ public class GameUpdateTask implements Runnable {
 
     private static final int DEFAULT_MAX_RETRIES = 5;
 
-    private Consumer<Exception> onError;
-    private InProgressGame game;
+    @Getter private InProgressGame game;
 
     @Setter private ScheduledFuture self;
-    @Setter private IntConsumer onChunkPulled;
-    @Setter private IntConsumer onKeyframePulled;
-    @Setter private Callback onFinished;
+    private IntConsumer onChunkPulled;
+    private IntConsumer onKeyframePulled;
+    private Consumer<Exception> onError;
+    private Callback onFinished;
+    private IntConsumer onKeyframeFailed;
+    private IntConsumer onChunkFailed;
 
     private boolean firstRun = true;
     private Map<Integer, Integer> retriesChunks = new HashMap<>();
@@ -63,6 +66,50 @@ public class GameUpdateTask implements Runnable {
     public GameUpdateTask(InProgressGame game, Consumer<Exception> errorCallback) {
         this.game = game;
         this.onError = errorCallback;
+    }
+
+    public void addOnChunkPulled(IntConsumer consumer) {
+        this.onChunkPulled = onChunkPulled == null ? consumer : onChunkPulled.andThen(consumer);
+    }
+
+    public void addOnKeyframePulled(IntConsumer consumer) {
+        this.onKeyframePulled = onKeyframePulled == null ? consumer : onKeyframePulled.andThen(consumer);
+    }
+
+    public void addOnError(Consumer<Exception> consumer) {
+        this.onError = onError == null ? consumer : onError.andThen(consumer);
+    }
+
+    public void addOnFinished(Callback callback) {
+        this.onFinished = callback.and(onFinished);
+    }
+
+    public void addOnKeyframeFailed(IntConsumer consumer) {
+        this.onKeyframeFailed = onKeyframeFailed == null ? consumer : onKeyframeFailed.andThen(consumer);
+    }
+
+    public void addOnChunkFailed(IntConsumer consumer) {
+        this.onChunkFailed = onChunkFailed == null ? consumer : onChunkFailed.andThen(consumer);
+    }
+
+    @Deprecated
+    public void setOnError(Consumer<Exception> consumer) {
+        onError = consumer;
+    }
+
+    @Deprecated
+    public void setOnFinished(Callback callback) {
+        onFinished = callback;
+    }
+
+    @Deprecated
+    public void setOnChunkPulled(IntConsumer consumer) {
+        this.onChunkPulled = consumer;
+    }
+
+    @Deprecated
+    public void setOnKeyframePulled(IntConsumer consumer) {
+        this.onKeyframePulled = consumer;
     }
 
     @Override
@@ -121,7 +168,7 @@ public class GameUpdateTask implements Runnable {
     }
 
     private void pullNewChunks(int maxId) {
-        log.debug("[" + game.getGameId() + "] " + (maxId - game.getLastAvailableChunk()) + " new chunks (" + (game.getLastAvailableChunk() + 1) + " to " + maxId + ")");
+        //log.debug("[" + game.getGameId() + "] " + (maxId - game.getLastAvailableChunk()) + " new chunks (" + (game.getLastAvailableChunk() + 1) + " to " + maxId + ")");
 
         // Time between chunks is occasionally less than the chunk time interval
         // So we check for all missed chunks here
@@ -138,8 +185,10 @@ public class GameUpdateTask implements Runnable {
         }
 
         if (!retriesExceeded.isEmpty()) {
-            retriesExceeded.forEach(retriesKeyframes::remove);
-            throw new RequestException("Unreachable keyframe: " + retriesExceeded);
+            for (int i: retriesExceeded) {
+                retriesKeyframes.remove(i);
+                if (onKeyframePulled != null) { onKeyframeFailed.accept(i); }
+            }
         }
 
         if (!retrySuccessful.isEmpty()) {
@@ -155,8 +204,10 @@ public class GameUpdateTask implements Runnable {
         }
 
         if (!retriesExceeded.isEmpty()) {
-            retriesExceeded.forEach(retriesChunks::remove);
-            throw new RequestException("Unreachable chunk: " + retriesExceeded);
+            for (int i: retriesExceeded) {
+                retriesChunks.remove(i);
+                if (onChunkFailed != null) { onChunkFailed.accept(i); }
+            }
         }
 
         if (!retrySuccessful.isEmpty()) {